requirements.txt
websocket-client = '1.8.0'
apache-airflow = ''
python-dotenv = '1.0.1'
pandas = '2.2.3'